requirements.txt
fastapi
uvicorn
jinja2
pyserial
websockets